# Conflicts:
#	FrontEnd/src/app/file-uploader/file-uploader.component.ts

+++ FrontEnd/src/app/file-uploader/file-uploader.component.ts
@@ -7,6 +7,8 @@ import {environment} from '../../environments/environment';
 import {ActivatedRoute, Params} from '@angular/router';
 import {Subscription} from 'rxjs';
 import {type} from 'os';
+import {ConfirmationDialogComponent} from '../confirmation-dialog/confirmation-dialog.component';
+import {MatDialog, MatDialogRef} from '@angular/material/dialog';
 
 @Component({
   selector: 'app-file-uploader',
@@ -22,10 +24,12 @@ export class FileUploaderComponent implements OnInit, OnDestroy {
   uploadSubscriptionDict$: {};
   subscription$: Subscription;
   public activeUploads;
+  private dialogRef: MatDialogRef<ConfirmationDialogComponent>;
 
   constructor(private api: ApiService,
               private messagesService: MessagesService,
-              private route: ActivatedRoute) {
+              private route: ActivatedRoute,
+              private dialog: MatDialog) {
     this.showCategoryLink = true;
     this.uploadSubscriptionDict$ = {};
     this.activeUploads = 0;
@@ -180,11 +184,22 @@ export class FileUploaderComponent implements OnInit, OnDestroy {
   }
 
   deleteFile(fileName: string): void {
-    if (fileName in this.uploadSubscriptionDict$) {
-      this.uploadSubscriptionDict$[fileName].unsubscribe();
-      this.api.deleteFile(fileName, true);
-    } else {
-      this.api.deleteFile(fileName);
-    }
+
+    this.dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      disableClose: false
+    });
+    this.dialogRef.componentInstance.confirmMessage = `Are you sure you want to delete the file '${fileName}'?`;
+
+    this.dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        if (fileName in this.uploadSubscriptionDict$) {
+          this.uploadSubscriptionDict$[fileName].unsubscribe();
+          this.api.deleteFile(fileName, true);
+        } else {
+          this.api.deleteFile(fileName);
+        }
+      }
+      this.dialogRef = null;
+    });
   }
 }
